% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bargpu}
\alias{bargpu}
\title{Bar Chart with ChartGPU}
\usage{
bargpu(data, mapping = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
Bar Chart with ChartGPU
}
\examples{
library(rchartgpu)

bargpu(
  ggplot2::economics[, c("date", "unemploy")],
  height = "400px"
)
}
