% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{chartgpu-shiny}
\alias{chartgpu-shiny}
\alias{chartgpuOutput}
\alias{renderChartgpu}
\title{Shiny bindings for chartgpu}
\usage{
chartgpuOutput(outputId, width = "100\%", height = "400px")

renderChartgpu(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates an HTML widget (or a
\href{https://rstudio.github.io/promises/}{promise} of an HTML widget).}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using \code{\link[=chartgpu]{chartgpu()}} within Shiny
applications and interactive Rmd documents.
}
\examples{

library(shiny)
library(rchartgpu)

ui <- fluidPage(
  fluidRow(
    column(
      width = 8,
      offset = 2,
      selectInput(
        inputId = "variable",
        label = "Variable:",
        choices = names(ggplot2::economics)[-1]
      ),
      chartgpuOutput(outputId = "chart", height = "500px")
    )
  )
)

server <- function(input, output, session) {

  output$chart <- renderChartgpu({
    chartgpu(ggplot2::economics[, c("date", input$variable)]) |>
      chartgpu_theme_light()
  })

}

if (interactive())
  shinyApp(ui, server)
}
