% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chartgpu-theme}
\alias{chartgpu-theme}
\alias{chartgpu_theme}
\alias{chartgpu_theme_light}
\alias{chartgpu_theme_dark}
\title{ChartGPU Theme}
\usage{
chartgpu_theme(chart, ...)

chartgpu_theme_light(chart)

chartgpu_theme_dark(chart)
}
\arguments{
\item{chart}{A chart created with \code{\link[=chartgpu]{chartgpu()}}.}

\item{...}{Options for theme, see \url{https://github.com/ChartGPU/ChartGPU/blob/main/docs/api/themes.md}.}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
ChartGPU Theme
}
\examples{

# Use built-in light theme
chartgpu(ggplot2::economics[, c("date", "psavert")]) |>
  chartgpu_theme_light()

# Use built-in light theme
chartgpu(ggplot2::economics[, c("date", "psavert")]) |>
  chartgpu_theme_dark()

# Custom theme
chartgpu(ggplot2::economics[, c("date", "psavert")]) |>
  chartgpu_theme(
    backgroundColor = "#002240",
    textColor = "#ed7e1b",
    axisLineColor = "#ed7e1b",
    axisTickColor = "#ed7e1b",
    gridLineColor = "#ed7e1b",
    colorPalette = list("#FFFFFF")
  )
}
