% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chartgpu_axis_x}
\alias{chartgpu_axis_x}
\title{X-Axis Options}
\usage{
chartgpu_axis_x(chart, ...)
}
\arguments{
\item{chart}{A chart created with \code{\link[=chartgpu]{chartgpu()}}.}

\item{...}{Options for x-axis,
see \url{https://github.com/ChartGPU/ChartGPU/blob/main/docs/api/options.md#axis-configuration}.}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
X-Axis Options
}
\examples{

# Provide name for axis
chartgpu(ggplot2::economics, aes(date, psavert)) |>
  chartgpu_axis_x(name = "Date")

# Custom date locale for axis labels
chartgpu(ggplot2::economics, aes(date, psavert)) |>
  chartgpu_axis_x(
    tickFormatter = JS("(ms) => new Date(ms).toLocaleDateString('de-DE')")
  )

}
