% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chartgpu_axis_y}
\alias{chartgpu_axis_y}
\title{Y-Axis Options}
\usage{
chartgpu_axis_y(chart, ...)
}
\arguments{
\item{chart}{A chart created with \code{\link[=chartgpu]{chartgpu()}}.}

\item{...}{Options for y-axis,
see \url{https://github.com/ChartGPU/ChartGPU/blob/main/docs/api/options.md#axis-configuration}.}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
Y-Axis Options
}
\examples{

# Provide name for axis
chartgpu(ggplot2::economics, aes(date, psavert)) |>
  chartgpu_axis_y(name = "personal savings rate")

# Custom labels for ticks + axis minimum/maximum
chartgpu(ggplot2::economics, aes(date, psavert)) |>
  chartgpu_axis_y(
    min = 0,
    max = 20,
    tickFormatter = JS("(v) => `${v.toFixed(0)} \%`")
  )
}
