% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chartgpu_options}
\alias{chartgpu_options}
\title{Update chart options}
\usage{
chartgpu_options(chart, ...)
}
\arguments{
\item{chart}{A chart created with \code{\link[=chartgpu]{chartgpu()}}.}

\item{...}{Any ChartGPU valid options,
see \url{https://github.com/ChartGPU/ChartGPU/blob/main/docs/api/options.md}}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
Update chart options
}
\examples{
# Specify any options

chartgpu(ggplot2::economics, aes(date, psavert)) |>
  chartgpu_options(
    animation = FALSE,
    yAxis = list(name = "personal savings rate")
  )

scattergpu(cars) |>
  chartgpu_options(
    theme = "light",
    tooltip = list(trigger = "axis")
  )
}
