% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chartgpu_serie}
\alias{chartgpu_serie}
\title{Update configuration of series}
\usage{
chartgpu_serie(chart, name, label = NULL, color = NULL, ...)
}
\arguments{
\item{chart}{A chart created with \code{\link[=chartgpu]{chartgpu()}}.}

\item{name}{Name of the serie (name of the variable in the data used)}

\item{label}{Label for display.}

\item{color}{Color for the serie.}

\item{...}{Other options.}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
Update configuration of series
}
\examples{

# Modify series options
data.frame(
  x = seq(-3, 7, by = 1/8),
  sin = sin(seq(-3, 7, by = 1/8)),
  cos = cos(seq(-3, 7, by = 1/8))
) |>
  chartgpu() |>
  chartgpu_serie(
    name = "sin",
    label = "Sine",
    color = "#228B22"
  ) |>
  chartgpu_serie(
    name = "cos",
    label = "Cosine",
    color = "#b22222",
    lineStyle = list(width = 4)
  )
}
