% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{chartgpu_zoom}
\alias{chartgpu_zoom}
\title{ChartGPU Data Zoom}
\usage{
chartgpu_zoom(chart, type = c("inside", "slider"), ...)
}
\arguments{
\item{chart}{A chart created with \code{\link[=chartgpu]{chartgpu()}}.}

\item{type}{Type of zoom between : \code{"inside"} (with mouse) or \code{"slider"} (below chart) or both.}

\item{...}{Options for zoom,
see \url{https://github.com/ChartGPU/ChartGPU/blob/main/docs/api/options.md#data-zoom-configuration}.}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
ChartGPU Data Zoom
}
\examples{

# Activate both zoom methods: mouse and slider
chartgpu(ggplot2::economics[, c("date", "psavert")]) |>
  chartgpu_zoom()

# Only slider
chartgpu(ggplot2::economics[, c("date", "psavert")]) |>
  chartgpu_zoom(type = "slider")
}
