% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{scattergpu}
\alias{scattergpu}
\title{Scatter Plot with ChartGPU}
\usage{
scattergpu(data, mapping = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{mapping}{Default list of aesthetic mappings to use for chart.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
Scatter Plot with ChartGPU
}
\examples{

# With a data.frame, 1st variable on X-axis, 2nd on Y-axis
scattergpu(cars)

# Provide mapping to select variable to plot
scattergpu(ggplot2::mpg, aes(cty, hwy))

# Color point based on variable
scattergpu(
  data = ggplot2::diamonds,
  mapping = aes(x = carat, y = price, color = clarity)
) |>
  chartgpu_theme_light(palette = scales::viridis_pal(option = "viridis")(8))
}
