% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartgpu.R
\name{chartgpu}
\alias{chartgpu}
\title{ChartGPU}
\usage{
chartgpu(
  data = NULL,
  type = NULL,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Can be a \code{data.frame} (first column is used on x-axis, the others on y-axis) or a list of parameters for configuring a chart.}

\item{type}{Type of chart, if \code{data} is a \code{data.frame}.}

\item{...}{Additional parameters.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\value{
A \code{\link[=chartgpu]{chartgpu()}} \code{htmlwidget} object.
}
\description{
ChartGPU is a TypeScript charting library built on WebGPU for smooth, interactive rendering—especially when you have lots of data. See https://github.com/ChartGPU/ChartGPU.
}
\examples{

# With a data.frame :
chartgpu(ggplot2::economics) # default to line chart

# scatter chart :
chartgpu(ggplot2::diamonds[, c("carat", "price")], type = "scatter")

# With a list passed directly to Javascript method :
chartgpu(list(
  series = list(
    list(
      type = "line",
      name = "Sine",
      data = data.frame(
        x = seq(-3, 7, by = 1/8),
        y = sin(seq(-3, 7, by = 1/8))
      )
    )
  )
))

# Two series and some options :
chartgpu(list(
  series = list(
    list(
      type = "line",
      name = "Sine",
      data = data.frame(
        x = seq(-3, 7, by = 1/8),
        y = sin(seq(-3, 7, by = 1/8))
      ),
      color = "#228B22" # forestgreen, use hex code
    ),
    list(
      type = "line",
      name = "Cosine",
      data = data.frame(
        x = seq(-3, 7, by = 1/8),
        y = cos(seq(-3, 7, by = 1/8))
      ),
      color = "#b22222" # firebrick, use hex code
    )
  ),
  xAxis = list(name = "X axis"),
  yAxis = list(name = "Y axis"),
  tooltip = list(trigger = "axis"),
  dataZoom = list(list(type = "inside"), list(type = "slider"))
))
}
